Separated the public ClientOptions in preparation of refactoring suggested by Corey.
